Patch from John Tytgat.
I've been using Sunfish sources to test our GCCSDK 4 code after I
added network support for the SCL runtime library case (instead of
using TCPIPLibs which is AOF and not ELF).
This resulted in a couple of harmless changes to allow both the
TCPIPLibs library to be used as well the network code in GCCSDK 4 SCL.
(*) Move include of socklib.h and unixlib.h behind a USE_TCPIPLIBS test
    and also add an include of riscos.h (which redefines strerror()
    into xstrerror()).
(*) Changed all xstrerror() calls in to strerror().
(*) Use size_t instead of unsigned int (triggered by the fact that
    iconv() uses size_t, not unsigned int)
    -> this avoids warnings
(*) Use socklen_t
    -> this avoids warnings
This allows me to build Sunfish (and Moonfish) module using the GCCSDK
cross compiler.

+++ src/imageentry_common.h
@@ -58,7 +58,7 @@ struct objinfo {
 ENTRYFUNCDECL(os_error *, gen_nfsstatus_error, (nstat stat));
 
 /* Convert a leafname into an nfs handle, following symlinks as necessary */
-ENTRYFUNCDECL(os_error *, leafname_to_finfo, (char *leafname, unsigned int *len, int simple, int followsymlinks, struct commonfh *dirhandle, struct objinfo **finfo, nstat *status, struct conn_info *conn));
+ENTRYFUNCDECL(os_error *, leafname_to_finfo, (char *leafname, size_t *len, int simple, int followsymlinks, struct commonfh *dirhandle, struct objinfo **finfo, nstat *status, struct conn_info *conn));
 
 /* Convert a full filename/dirname into an nfs handle */
 ENTRYFUNCDECL(os_error *, filename_to_finfo, (char *filename, int followsymlinks, struct objinfo **dinfo, struct objinfo **finfo, char **leafname, int *filetype, int *extfound, struct conn_info *conn));
