+++ src/nfs3-procs.c
@@ -49,13 +49,13 @@ static nstat diropargs_to_path(struct diropargs3 *where, char **path, int *filet
 	char *dirpath;
 	static char buffer[MAX_PATHNAME];
 	char *leaf = where->name.data;
-	unsigned leaflen = where->name.size;
+	size_t leaflen = where->name.size;
 
 	NR(nfs3fh_to_path(&(where->dir), &dirpath, conn));
 
 	if (choices.fromenc != (iconv_t)-1) {
 		char *encleaf;
-		unsigned encleaflen;
+		size_t encleaflen;
 		static char buffer2[MAX_PATHNAME];
 
 		encleaf = buffer2;
@@ -584,7 +584,7 @@ enum accept_stat NFSPROC3_READDIR(struct readdirargs3 *args, struct readdirres3 
 		if (read > 0) {
 			struct entry3 *entry;
 			char *leaf = buffer + 20;
-			unsigned int leaflen;
+			size_t leaflen;
 			int filetype;
 			unsigned int load = ((unsigned int *)buffer)[0];
 			int type;
@@ -605,7 +605,7 @@ enum accept_stat NFSPROC3_READDIR(struct readdirargs3 *args, struct readdirres3 
 
 			if (choices.toenc != (iconv_t)-1) {
 				char *encleaf;
-				unsigned encleaflen;
+				size_t encleaflen;
 				static char buffer2[MAX_PATHNAME];
 
 				encleaf = buffer2;
@@ -682,7 +682,7 @@ enum accept_stat NFSPROC3_READDIRPLUS(struct readdirplusargs3 *args, struct read
 		if (read > 0) {
 			struct entryplus3 *entry;
 			char *leaf = buffer + 20;
-			unsigned int leaflen;
+			size_t leaflen;
 			int filetype;
 			unsigned int load = ((unsigned int *)buffer)[0];
 			unsigned int exec = ((unsigned int *)buffer)[1];
@@ -708,7 +708,7 @@ enum accept_stat NFSPROC3_READDIRPLUS(struct readdirplusargs3 *args, struct read
 
 			if (choices.toenc != (iconv_t)-1) {
 				char *encleaf;
-				unsigned encleaflen;
+				size_t encleaflen;
 				static char buffer2[MAX_PATHNAME];
 
 				encleaf = buffer2;
